Add vault v2 namespace and move files to it
