+++ modules/ppcp-vaulting/src/CustomerApprovalListener.php
@@ -11,8 +11,8 @@ namespace WooCommerce\PayPalCommerce\Vaulting;
 
 use Exception;
 use Psr\Log\LoggerInterface;
-use WooCommerce\PayPalCommerce\ApiClient\Endpoint\PaymentTokenEndpoint;
 use WooCommerce\PayPalCommerce\ApiClient\Exception\AlreadyVaultedException;
+use WooCommerce\PayPalCommerce\ApiClient\VaultV2\PaymentTokenEndpoint;
 use WooCommerce\PayPalCommerce\WcSubscriptions\FreeTrialHandlerTrait;
 
 /**
